+++ examples/03-dev/demo/project/src/js/test-2/index.js
@@ -1,4 +1,5 @@
 
+require('bootstrap/dist/css/bootstrap.css');
 require('../../css/test/index.css');
 require('../../css/test/index2.css');
 require('../../css/test/index.less');
@@ -13,3 +14,4 @@ console.log('test-2/index');
 
 // test ignoreNodeModules
 var Swiper = require('swiper');
+require('bootstrap');