+++ examples/03-dev/demo/project/src/js/test/index.js
@@ -1,4 +1,5 @@
 
+require('bootstrap/dist/css/bootstrap.css');
 require('../../css/test/index.css');
 require('../../css/test/index2.css');
 require('../../css/test/index.less');