+++ examples/03-dev/dist-cus.js
@@ -1,0 +1,6 @@
+
+require('../../util/change_cwd_to')(__dirname + '/demo');
+
+require('./cus');
+
+require('../../util/exec')('lila dist test-3/index');